+++ packages/cli/bin/tidypress.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
-import { runCliMain } from '../dist/index.js'
+import { runCliMain } from '../dist/runCli.js'
 
 await runCliMain(process.argv.slice(2))